+++ packages/services/f-analytics/src/utils/helpers.js
@@ -1,7 +1,9 @@
 import { windowServices } from '@justeat/f-services';
 import { GROUPS_FEATURES_ROUTES } from '../constants';
 
-// TODO: jsdoc
+/**
+ * Returns string description based on window width
+ */
 export const getDisplaySize = () => {
     const width = windowServices.getWindowWidth();
 
@@ -13,7 +15,9 @@ export const getDisplaySize = () => {
     return 'narrow';
 };
 
-// TODO: jsdoc
+/**
+ * Returns string Portrait|Landscape based on window width/height
+ */
 export const getOrientation = () => {
     const height = windowServices.getWindowHeight();
     const width = windowServices.getWindowWidth();
@@ -21,10 +25,22 @@ export const getOrientation = () => {
     return height > width ? 'Portrait' : 'Landscape';
 };
 
+/**
+ * Returns array, filtered when a feature invludes the route param
+ * @param {String} route
+ */
 const getGroupObjectFromRoute = route => GROUPS_FEATURES_ROUTES.filter(groups => groups.features.find(x => x.routes.includes(route)));
 
+/**
+ * Returns group name
+ * @param {String} route
+ */
 export const mapRouteToGroup = route => getGroupObjectFromRoute(route).map(x => x.group).toString() || route;
 
+/**
+ * Returns page name
+ * @param {String} route
+ */
 export const mapRouteToFeature = route => {
     const feature = getGroupObjectFromRoute(route).map(group => group.features.filter(x => x.routes.includes(route)).map(x => x.feature)).toString();
     return feature || route;