+++ packages/services/f-analytics/CHANGELOG.md
@@ -3,6 +3,23 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.7.0
+------------------------------
+*August 9, 2021*
+
+### Added
+- `PreparePageData` method
+
+### Changed
+- Replaced `MAP_ROUTE_TO_FEATURE_NAME` const with `mapRouteToGroup` & `mapRouteToFeature`
+
+
+
+v0.6.0
+------------------------------
+*August 9, 2021*
+
+
 v0.5.0
 ------------------------------
 *July 28, 2021*